workaround for tests
