+++ src/Common/DistributedTracing.h
@@ -99,26 +99,11 @@ private:
 
 struct IDistributedTracer;
 
-std::shared_ptr<Span> getCurrentSpan();
-std::shared_ptr<IDistributedTracer> getCurrentTracer();
+//std::shared_ptr<Span> switchToChildSpan(const String& operation_name);
 
-std::shared_ptr<Span> switchToChildSpan(const String& operation_name);
-
-void switchToParentSpan(const std::shared_ptr<Span>& parent_span);
+//void switchToParentSpan(const std::shared_ptr<Span>& parent_span);
 
 std::shared_ptr<Span> parkSpan(const std::shared_ptr<Span>& new_span);
-
-class SpanGuard
-{
-public:
-    SpanGuard();
-    SpanGuard(const String& operation_name);
-    ~SpanGuard();
-
-private:
-    std::shared_ptr<DB::opentracing::Span> parent_span;
-};
-
 
 class ParkingSpanGuard
 {
