+++ src/IO/ReadBufferAIO.cpp
@@ -2,6 +2,7 @@
 
 #include <IO/ReadBufferAIO.h>
 #include <IO/AIOContextPool.h>
+#include <Common/DistributedTracing.h>
 #include <Common/ProfileEvents.h>
 #include <Common/Stopwatch.h>
 #include <Core/Defines.h>
@@ -202,6 +203,8 @@ off_t ReadBufferAIO::seek(off_t off, int whence)
 void ReadBufferAIO::synchronousRead()
 {
     CurrentMetrics::Increment metric_increment_read{CurrentMetrics::Read, 1, CurrentMetrics::TracingMode::COMPLETE};
+//    opentracing::SpanGuard span_guard("synchronousRead");
+    opentracing::Span span("synchronousRead", {}, 0);
 
     prepare();
     bytes_read = ::pread(fd, buffer_begin, region_aligned_size, region_aligned_begin);
@@ -238,6 +241,8 @@ bool ReadBufferAIO::waitForAIOCompletion()
         return false;
 
     CurrentMetrics::Increment metric_increment_read{CurrentMetrics::Read, 1, CurrentMetrics::TracingMode::COMPLETE};
+//    opentracing::SpanGuard span_guard("waitForAIOCompletion");
+    opentracing::Span span("waitForAIOCompletion", {}, 1);
 
     bytes_read = future_bytes_read.get();
     is_pending_read = false;
