+++ src/IO/ReadBufferFromFileDescriptor.cpp
@@ -5,6 +5,7 @@
 #include <Common/Stopwatch.h>
 #include <Common/Exception.h>
 #include <Common/CurrentMetrics.h>
+#include <Common/DistributedTracing.h>
 #include <IO/ReadBufferFromFileDescriptor.h>
 #include <IO/WriteHelpers.h>
 
@@ -53,6 +54,9 @@ bool ReadBufferFromFileDescriptor::nextImpl()
         ssize_t res = 0;
         {
             CurrentMetrics::Increment metric_increment{CurrentMetrics::Read, 1, CurrentMetrics::TracingMode::COMPLETE};
+            opentracing::Span span("ReadBufferFromFileDescriptor", {}, 2);
+//            opentracing::SpanGuard span_guard("ReadBufferFromFileDescriptor::nextImpl");
+
             res = ::read(fd, internal_buffer.begin(), internal_buffer.size());
         }
         if (!res)
